+++ NeKzBot/Server/Events.cs
@@ -46,7 +46,7 @@ namespace NeKzBot.Server
 		{
 			try
 			{
-				await (await Utils.FindTextChannel(Configuration.Default.LogChannelName))?.SendMessage($"**{await Utils.GetLocalTime()}**\n{Bot.Client.CurrentUser?.Name}#{Bot.Client.CurrentUser?.Discriminator.ToString("D4")} joined a server.\nName • {e.Server?.Name} (ID {e.Server?.Id})\nOwner • {(e.Server?.Owner?.Name != null ? $"*{e.Server?.Owner.Name}#{e.Server?.Owner.Discriminator.ToString("D4")}*" : "*not found*")} (ID {e.Server?.Owner.Id})");
+				await (await Utils.FindTextChannel(Configuration.Default.LogChannelName))?.SendMessage($"**{await Utils.GetLocalTimeAsync()}**\n{Bot.Client.CurrentUser?.Name}#{Bot.Client.CurrentUser?.Discriminator.ToString("D4")} joined a server.\nName • {e.Server?.Name} (ID {e.Server?.Id})\nOwner • {(e.Server?.Owner?.Name != null ? $"*{e.Server?.Owner.Name}#{e.Server?.Owner.Discriminator.ToString("D4")}*" : "*not found*")} (ID {e.Server?.Owner.Id})");
 			}
 			catch (Exception ex)
 			{
@@ -59,7 +59,7 @@ namespace NeKzBot.Server
 		{
 			try
 			{
-				await (await Utils.FindTextChannel(Configuration.Default.LogChannelName))?.SendMessage($"**{await Utils.GetLocalTime()}**\n{Bot.Client.CurrentUser?.Name}#{Bot.Client.CurrentUser?.Discriminator.ToString("D4")} left a server.\nName • {e.Server?.Name} (ID {e.Server?.Id})\nOwner • {(e.Server?.Owner?.Name != null ? $"*{e.Server?.Owner.Name}#{e.Server?.Owner.Discriminator.ToString("D4")}*" : "*not found*")} (ID {e.Server?.Owner.Id})");
+				await (await Utils.FindTextChannel(Configuration.Default.LogChannelName))?.SendMessage($"**{await Utils.GetLocalTimeAsync()}**\n{Bot.Client.CurrentUser?.Name}#{Bot.Client.CurrentUser?.Discriminator.ToString("D4")} left a server.\nName • {e.Server?.Name} (ID {e.Server?.Id})\nOwner • {(e.Server?.Owner?.Name != null ? $"*{e.Server?.Owner.Name}#{e.Server?.Owner.Discriminator.ToString("D4")}*" : "*not found*")} (ID {e.Server?.Owner.Id})");
 			}
 			catch (Exception ex)
 			{
@@ -72,7 +72,7 @@ namespace NeKzBot.Server
 		{
 			try
 			{
-				await (await Utils.FindTextChannel(Configuration.Default.LogChannelName, e.Server))?.SendMessage($"**{await Utils.GetLocalTime()}**\n{e.User?.Name}#{e.User?.Discriminator.ToString("D4")} (ID {e.User?.Id}) has been banned from the server.");
+				await (await Utils.FindTextChannel(Configuration.Default.LogChannelName, e.Server))?.SendMessage($"**{await Utils.GetLocalTimeAsync()}**\n{e.User?.Name}#{e.User?.Discriminator.ToString("D4")} (ID {e.User?.Id}) has been banned from the server.");
 			}
 			catch (Exception ex)
 			{
@@ -85,7 +85,7 @@ namespace NeKzBot.Server
 		{
 			try
 			{
-				await (await Utils.FindTextChannel(Configuration.Default.LogChannelName, e.Server))?.SendMessage($"**{await Utils.GetLocalTime()}**\n{e.User?.Name}#{e.User?.Discriminator.ToString("D4")} (ID {e.User?.Id}) has been unbanned from the server.");
+				await (await Utils.FindTextChannel(Configuration.Default.LogChannelName, e.Server))?.SendMessage($"**{await Utils.GetLocalTimeAsync()}**\n{e.User?.Name}#{e.User?.Discriminator.ToString("D4")} (ID {e.User?.Id}) has been unbanned from the server.");
 			}
 			catch (Exception ex)
 			{
@@ -98,7 +98,7 @@ namespace NeKzBot.Server
 		{
 			try
 			{
-				await (await Utils.FindTextChannel(Configuration.Default.LogChannelName, e.Server))?.SendMessage($"**{await Utils.GetLocalTime()}**\n{e.User?.Name}#{e.User?.Discriminator.ToString("D4")} (ID {e.User?.Id}) joined the server.");
+				await (await Utils.FindTextChannel(Configuration.Default.LogChannelName, e.Server))?.SendMessage($"**{await Utils.GetLocalTimeAsync()}**\n{e.User?.Name}#{e.User?.Discriminator.ToString("D4")} (ID {e.User?.Id}) joined the server.");
 
 				// Give new user a world record role if he has one on speedrun.com or on board.iverb.me
 				if (!((bool)(e?.User?.IsBot))
@@ -130,7 +130,7 @@ namespace NeKzBot.Server
 						}
 						else
 							return;
-						await (await Utils.FindTextChannel(Configuration.Default.LogChannelName, e.Server))?.SendMessage($"**{await Utils.GetLocalTime()}**\n{e.User?.Name}#{e.User?.Discriminator.ToString("D4")} (ID {e.User?.Id}) has earned the _{Configuration.Default.WorldRecordRoleName}_ role.");
+						await (await Utils.FindTextChannel(Configuration.Default.LogChannelName, e.Server))?.SendMessage($"**{await Utils.GetLocalTimeAsync()}**\n{e.User?.Name}#{e.User?.Discriminator.ToString("D4")} (ID {e.User?.Id}) has earned the _{Configuration.Default.WorldRecordRoleName}_ role.");
 					}
 				}
 			}
@@ -145,7 +145,7 @@ namespace NeKzBot.Server
 		{
 			try
 			{
-				await (await Utils.FindTextChannel(Configuration.Default.LogChannelName, e.Server))?.SendMessage($"**{await Utils.GetLocalTime()}**\n{e.User?.Name}#{e.User?.Discriminator.ToString("D4")} (ID {e.User?.Id}) left or was kicked from the server.");
+				await (await Utils.FindTextChannel(Configuration.Default.LogChannelName, e.Server))?.SendMessage($"**{await Utils.GetLocalTimeAsync()}**\n{e.User?.Name}#{e.User?.Discriminator.ToString("D4")} (ID {e.User?.Id}) left or was kicked from the server.");
 			}
 			catch (Exception ex)
 			{
@@ -183,7 +183,7 @@ namespace NeKzBot.Server
 							await e.After.AddRoles(role);
 							// Not adding this to the streaming list, somebody could have abused this by changing his game url multiple times
 							//var result = await Utils.AddDataAsync(index, channel);
-							await (await Utils.FindTextChannel(Configuration.Default.LogChannelName, e.Server))?.SendMessage($"**{await Utils.GetLocalTime()}**\n{e.After.Name}#{e.After.Discriminator.ToString("D4")} (ID {e.After.Id}) has  earned the _{Configuration.Default.StreamingRoleName}_ role.");
+							await (await Utils.FindTextChannel(Configuration.Default.LogChannelName, e.Server))?.SendMessage($"**{await Utils.GetLocalTimeAsync()}**\n{e.After.Name}#{e.After.Discriminator.ToString("D4")} (ID {e.After.Id}) has  earned the _{Configuration.Default.StreamingRoleName}_ role.");
 						}
 					}
 				}