+++ NeKzBot/Utilities/RestUtils.cs
@@ -14,19 +14,18 @@ namespace NeKzBot.Utilities
 		public static Task<bool> ValidPathName(string path)
 			=> Task.FromResult((!(string.IsNullOrEmpty(path))) && (path?.IndexOfAny(Path.GetInvalidPathChars()) == -1));
 
-		public static Task<string> GetLocalTime(bool actualtime = false)
+		public static async Task<string> GetLocalTimeAsync(bool actualtime = false)
 		{
-			var result = default(string);
 			if (actualtime)
 			{
-				var zone = TimeZoneInfo.FindSystemTimeZoneById("Central European Standard Time");
+				var zone = (await IsLinux())
+									? TimeZoneInfo.FindSystemTimeZoneById("Europe/Vienna")
+									: TimeZoneInfo.FindSystemTimeZoneById("Central European Standard Time");
 				var offset = zone.GetUtcOffset(DateTime.UtcNow).Hours;
-				result = $"{TimeZoneInfo.ConvertTimeFromUtc(DateTime.UtcNow, zone).ToString("HH:mm:ss")}" +
-						 $" (UTC{((offset == 0) ? " " : (offset < 0) ? "-" : "+")}{offset})";
+				return $"{TimeZoneInfo.ConvertTimeFromUtc(DateTime.UtcNow, zone).ToString("HH:mm:ss")}" +
+					   $" (UTC{((offset == 0) ? " " : (offset < 0) ? "-" : "+")}{offset})";
 			}
-			else
-				result = DateTime.UtcNow.ToString("HH:mm:ss");
-			return Task.FromResult(result);
+			return DateTime.UtcNow.ToString("HH:mm:ss");
 		}
 
 		public static Task<string> GetAppPath()